--- src/main/java/com/devnity/devnity/domain/mapgakco/dto/mapgakcocomment/MapgakcoCommentCreateRequest.java
@@ -1,5 +1,0 @@
-package com.devnity.devnity.domain.mapgakco.dto.mapgakcocomment;
-
-public class MapgakcoCommentCreateRequest {
-  // temp
-}
